fix(player-ticket): rename files + reduce expiry to 4h
