+++ server/src/modules/player-ticket/player-ticket.consts.ts
@@ -1,0 +1,1 @@
+export const PLAYER_TICKET_EXPIRY = 4 * 60 * 60 * 1000;
